Implicit constants and free variables
Planning languages usually take the interrogation mark char as a way to
tell the parser the token is a variable. Now Hype will use this
internally and will consider all other tokens as constants. This is the
first step to deal with constant objects in the operators and methods
description. Parsers and compilers became simpler, without the need to
remove and add '?'. The only exception is the Hyper_Compiler that is
limited to use an a-z char at the beginning. No side-effect is expected
for the files alredy supported in older versions.