Positive and negative
Easier to think about positive and negative preconditions instead of
true and false, as a false goal may give the wrong idea about an actual
negative goal (something not desired but real).
True and false are more about the actual implementation than the
abstraction.
All variables and comments changed to match this expression.