Predicate partitioning
Ignoring useless cases (nil), which requires compiler modification to
match with useless goals from original input.
A useless predicate may appear in the initial state, but never in the
operators, or appear for the first time in the goal state (impossible to
achieve).
Reversed mutable (true) and constant (false) to simplify future
extensions.